Rev-X battery recharge overhauled
The old DECORATE-based battery recharge code for the Rev-X has been completely replaced with an ACS script. This change allows the Rev-X to recharge even when the player is not using it, though the recharge rate had to be increased by 3 tics to prevent infinite max. power with the Rapid Fire Coolant.